[ARM] msm: dma: Add msm_dmov_enqueue_cmd_no_callback and msm_dmov_add_callback
Use msm_dmov_enqueue_cmd_no_callback instead of msm_dmov_enqueue_cmd to
start a dma transfer without turning on the dma interrupt. This allows the
dma to be active while the arm11 is in power collapse.
Call msm_dmov_add_callback to enable the interrupt or call complete_func from
a tasklet if the dma has already finished.

Also, fix the interrupt handler to release the spinlock before calling
the callbacks.

Signed-off-by: Arve Hjønnevåg <arve@android.com>