Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

This needs some testing before it is submitted. If the hardware ecc for all 0 results in an all 0 checksum, this may be the right fix, but if not, then the the ecc check does not work, and should be fixed instead.

Patch-set: 1
Reviewer: Arve Hjønnevåg <1000205@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0