[ARM] msm: dal: Don't copy the client data when the waiter wakes up.
The smd handler could be in the middle of processing the next smd notification
when the dal call waiter wakes up. The waiter then tried to copy data out of the
temporary client data buffer, and was reading the new transaction in progress,
in this case an async notification. Since the dal smd notifier already copied
the correct data to the reply buffer, nothing needs to be done on wake but
to return the status.

Signed-off-by: Dima Zavin <dima@android.com>