UBIFS: fix GC error path
This patch fixes the GC error patch:

1. We should first synchronize the GC write-buffer, then switch to
   R/O mode, not vice-versa.
2. If we went out of the main loop with break, we already have
   lprops released, we cannot do 'goto out', because that code
   path returns the LEB, which should not be done.

Also, move the 'if (c->ro_media)' out of the mutex - in this case
we do not have to release the mutex if that is true, and we do not
need the 'out_unlock' label.

And finally, rename the 'out' label to 'out_err' to make it clear
that this is error path.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>