Add interrupt for oprofile on Nexus One
This patch adds a definition of the correct interrupt to use for the
performance monitoring hardware on the MSM72k CPU. With this patch
oprofile is able to collect profile samples on the Nexus One.

The patch also makes the kernel driver report an error if no interrupt
definition is available, instead of falsely reporting that profiling
has been successfully enabled.

Signed-off-by: Mike Playle <mike.playle@realvnc.com>