UBIFS: temporary patch to prepare for ubifs_scan changes
Currently 'ubifs_scan()' assumes that the scanned eraseblock does not contain
corruptions, and if it does, 'ubifs_scan()' frees all the scanned information
and returns -EUCLEAN.

'ubifs_scan()', however, has the 'quiet' argument which can prevent it from
printing many error messages about the corruption.

If we look at 'ubifs_scan()' users that pass 'quiet = 1', we can see that the
typical scenario is:

1. ubifs_scan(quiet = 1);
2. if error is -EUCLEAN then
3.    scan the media again and recover from errors

Thus, the users read the flash twice, which is unnecessary and which is slow.
And the intention is to

1. make 'ubifs_scan()' no free the scanned information in case of corruption
   if quiet is 1.
2. then the users do not have to scan for the second time.

But this requires a lot of changes. So we first introduce this temporary hack
where we do not free the scanned information if quiet is 2. When we modify
all the callers, we can remove this temporary hack.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>