UBIFS: add dbg_dump_sleb helper function
We have 'dbg_dump_leb()' function which accepts LEB number, then scans it and
dumps. However, sometimes we already have the scanning information and we do
not want to re-scan. This patch introduces a 'dbg_dump_sleb()' function which
takes the scanning information as an argument. Also, it changes
'dbg_dump_leb()' and makes it use 'dbg_dump_sleb()' internally.

Additionally, this patch switches one of the users to 'dbg_dump_sleb()'.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>