UBIFS: do not recover if master nodes offsets are bogus
This patch makes 'scan_for_master()' fail with a fatal error if we have the
same amount of valid master nodes in both master LEBs, but the last master
nodes are at different offsets. This cannot be caused by a power cut, thus
we should not try to recover from this error.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>