UBIFS: remove common message for unrecoverable errors
This is just a preparation patch which removes the error message from the
"out_unrecoverable:" label in 'scan_for_master()'. This change is need because
I am going to add more checks for unrecoverable errors, so that we'll need
different error messages for different cases. IOW, this is just a preparational
patch.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>