CHROMIUM: LSM: chromiumos security module
Initially the chromiumos security module only prevents symlinks
in mount paths. Future versions will be more restrictive and
will be configurable using a driver interface.

This is a squash of the following commits:
  CHROMIUM: kernel: Initial chromiumos security module.
  CHROMIUM: LSM: add module loading restrictions
  CHROMIUM: LSM: spinlock the root directory lock
  CHROMIUM: LSM: fix sysfs parameter exposure
  CHROMIUM: LSM: switch to stacked Chromium OS LSM
  CHROMIUM: LSM: handle pinned filesystem umount
  CHROMIUM: LSM: enable firmware pinning
  CHROMIUM: LSM: check vfsmount to avoid disengagement
  CHROMIUM: LSM: Assume read-only root if device is NULL
  CHROMIUM: Don't allow building LSM as a module
  CHROMIUM: LSM: kernel flag to allow mounting to symlinked path

Additionally, some fixups were applied to make the code work on top of
kernel 4.4 and the registration was updated to use regular LSM hooks.

Additional interfaces were exposed in fs/namei.c and do_mount() in
fs/namespace.c was updated to set up a temporary current->nameidata
while resolving the mountpoint path until sb_mount security hooks are
called, in order to make it possible to still check total_link_count
from those hooks.

It is still not possible to build LSMs as loadable modules, since
security_hooks_heads (in security/security.c) is not exported. This
could be fixed with the following upstream change:
  EXPORT_SYMBOL(security_hook_heads);
For now, we're happy to keep this as a built-in only.

BUG=chromium-os:21954
TEST=Use a mount path with a symlink and observe mount failure.

Change-Id: I47ade3b7be684ab9e0533ce5ffd1e81009eaebfe
Signed-off-by: Stephan Uphoff <ups@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/10581
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>

[rebase44(filbranden): Resolved trivial conflicts in Kconfig.  Updated
 Makefile to use a separate "chromiumos_lsm" name for the module.
 Introduced nameidata_get_total_link_count() and other APIs to make it
 possible to keep a valid nameidata and to access that field from this
 module.  Updated registration to use the LSM_HOOK_INIT API instead of
 adding explicit hooks to global security/security.c]
Signed-off-by: Filipe Brandenburger <filbranden@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/334277
Reviewed-on: https://chromium-review.googlesource.com/334278
Reviewed-on: https://chromium-review.googlesource.com/334279

Conflicts:
	security/Makefile

[rebase412(groeck):
	Resolved conflicts; include file name changes; API changes
Signed-off-by: Guenter Roeck <groeck@chromium.org>