CHROMIUM: memconsole: Add ACPI method to locate cbmem start
With >=4GB memory config the memory map can be split and relying
on total memory for a physical address to scan near does not work.

Instead of relying on scanning memory there is now an NVS field
defined in coreboot to get the pointer to the first cbmem entry.

The existing memory walking method is left in place (for now?)
with one fix to check the pointer returned by ioremap.

BUG=chrome-os-partner:7350
TEST=Tested with Stumpy 2GB and 4GB config:

1) install latest coreboot firmware
2) install kernel with this patch
03 verify that /sys/firmware/log is valid

Change-Id: I593fe03027b09d49da7dc133de7aab603c566208
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/13476
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>

[benzh: 3.14 rebase. Fix acpi header due to upstream change
8b48463f89429af408ff695244dc627e1acff4f7 ACPI: Clean up inclusions of ACPI header files
]
Signed-off-by: Ben Zhang <benzh@chromium.org>