CHROMIUM: dark resume: move to general platform api
This changes dark resume so that more control is given to the architecture
specific code to check what woke the system up. A callback that is registered
that checks if any of the devices in the source list woke the system. This
function can call into platform specific code to figure out what resumed the
system.

Since we don't know what a platform may require to check the wake source, this
also changes a spinlock used to protect the source list to a mutex. Platforms
such as x86 may use ACPI which can sleep when evaluating an ACPI object.

BUG=chromium:223578
TEST=manual on rambi
	'echo enabled > /sys/class/rtc/rtc0/device/power/dark_resume_source'
	'echo +10 > /sys/class/rtc/rtc0/wakealarm ; echo mem > /sys/power/state'
	check that 'cat /sys/power/dark_resume_state' is 1

Change-Id: If1afd3e6a333494b304feacfed879d85e7cfeb3b
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/58974
Reviewed-on: https://chromium-review.googlesource.com/191079
Reviewed-by: Sameer Nanda <snanda@chromium.org>

Conflicts:
	drivers/rtc/rtc-cmos.c

[rebase44(groeck): Dropped changes in drivers/rtc/rtc-cmos.c.]
Signed-off-by: Guenter Roeck <groeck@chromium.org>