net.h/skbuff.h: Remove extern from function prototypes
There are a mix of function prototypes with and without extern
in the kernel sources.  Standardize on not using extern for
function prototypes.

Function prototypes don't need to be written with extern.
extern is assumed by the compiler.  Its use is as unnecessary as
using auto to declare automatic/local variables in a block.

Signed-off-by: Joe Perches <joe@perches.com>
(cherry picked from commit 7965bd4d71ef7cf1db00afb9e406ddfc13443c13)
Signed-off-by: Alex Shi <alex.shi@linaro.org>

 Conflicts:
	include/linux/skbuff.h