ARM: remove name from machine_desc for DT platforms
As part of the effort to ultimately remove struct machine_desc, the
machine name is not really needed. It is only used for /proc/cpuinfo
"Hardware" field. Get this information from the DT instead and remove
the name string in the machine descriptor for DT machine descriptors. The
model or machine compatible property from the DT is used instead.

This likely changes the name reported by /proc/cpuinfo. There could be
some userspace dependencies on the old name. Debian libdebian-installer
used to use this to determine kernel flavour, but now uses the DT
directly. It is not really a good idea for userspace to depend on this
anyway.

Signed-off-by: Rob Herring <rob.herring@calxeda.com>