ARM: 8070/1: Introduce arm_get_current_stack_frame()
Currently there are numerous places where "struct pt_regs" are used to
populate "struct stackframe", however all of those location do not
consider the situation where the kernel might be compiled in THUMB2
mode, in which case the framepointer member of pt_regs become ARM_r7
instead of ARM_fp (r11). Document this idiosyncracy in the
definition of "struct stackframe"

The easiest solution is to introduce a new function (in the spirit of
https://groups.google.com/forum/#!topic/linux.kernel/dA2YuUcSpZ4)
which would hide the complexity of initializing the stackframe struct
from pt_regs.

Also implement a macro frame_pointer(regs) that would return the correct
register so that we can use it in cases where we just require the frame
pointer and not a whole struct stackframe

Change-Id: Ib4ba32c72e39c51e2d1f08ad144c5570caed05a1
Signed-off-by: Nikolay Borisov <Nikolay.Borisov@arm.com>
Acked-by: Will Deacon <will.deacon@arm.com>
Acked-by: Robert Richter <rric@kernel.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Git-commit: 9865f1d46a68a5f58cb623054c8308484007d8a4
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
[ymalik@codeaurora.org: resolve trivial merge conflicts]
Signed-off-by: Yasir Malik <ymalik@codeaurora.org>
CRs-Fixed: 2020623