ARM: expand fixmap region to 3MB
With commit a05e54c103b0b8 "ARM: 8031/2: change fixmap mapping region to
support 32 CPUs", the fixmap region was expanded to 2MB, but it
precluded any other uses of the fixmap region in order to support up to
32 CPUs. In order to support other uses the fixmap region needs to be
expanded beyond 2MB. Fortunately, the adjacent 1MB range
0xffe00000-0xfff00000 is availabe to use.

Remove fixmap_page_table ptr and lookup the page table via the virtual
address so that the fixmap region can span more that one pmd. The 2nd
pmd is already created since it is shared with the vector page.

Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Kees Cook <keescook@chromium.org>
Cc: Russell King <linux@arm.linux.org.uk>
Signed-off-by: Rob Herring <robh@kernel.org>