inet: frag: use seqlock for hash rebuild
rehash is rare operation, don't force readers to take
the read-side rwlock.

Instead, we only have to detect the (rare) case where
the secret was altered while we are trying to insert
a new inetfrag queue into the table.

If it was changed, drop the bucket lock and recompute
the hash to get the 'new' chain bucket that we have to
insert into.

Joint work with Nikolay Aleksandrov.

Signed-off-by: Florian Westphal <fw@strlen.de>
Signed-off-by: Nikolay Aleksandrov <nikolay@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit ab1c724f633080ed2e8a0cfe61654599b55cf8f9)
Signed-off-by: Alex Shi <alex.shi@linaro.org>

 Conflicts:
	net/ipv6/netfilter/nf_conntrack_reasm.c
	net/ipv6/reassembly.c