UPSTREAM: ARM: qcom: scm: Fix incorrect cache invalidation
The cache invalidation in scm_call() correctly rounds down the start
address to invalidate the beginning of the cacheline but doesn't properly
round up the 'end' address to make it aligned.  The last chunk of the
buffer won't be invalidated when 'end' is not cacheline size aligned so
make sure to invalidate the last few bytes in such situations. It also
doesn't do anything about outer caches so make sure to invalidate and flush
those as well.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Kumar Gala <galak@codeaurora.org>

BUG=chrome-os-partner:49255
TEST=Passed QCA non-ChromiumOS test suite

(cherry picked from commit f76c6916570d8d24a622f2107ef747531148046e)
Signed-off-by: Cristian Prundeanu <cprundea@codeaurora.org>
Signed-off-by: Sreedhar Sambangi <ssambang@codeaurora.org>
Change-Id: I0e96892c2087639e957ae6cce4f64e3518b050c5
Reviewed-on: https://chromium-review.googlesource.com/327741
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>