Update patch set 1
Patch Set 1:

This change was already merged in kernel/common (see:  Id0e1cff5828062009b4f94c987ac91f88f14652e ) Would be it be possible to have it landed on flo device as well? With this commit some keys that are currently ignored will be handled correctly by Android. I pushed it in android-msm-flo-3.4-kitkat-mr2 as it is the branch currently used for the prebuilt kernel for Android master branch in flo.

Note that we are also working in parallel (with Usunk from CTS team and MichaelW from framework) to have a CTS test to check the presence of this patch but currently the CTS framework doesn't permit to host such tests. It should be the case in the future and we are continuing to look to have a CTS test to check the presence of this patch.

Patch-set: 1
Reviewer: Olivier Gay <1017523@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0