fuse: sync release from flush
If FOPEN_SYNC_RELEASE flag is set from open, then try doing a synchronous
release from ->flush().

Set FUSE_RELEASE_ISSYNC flag in the RELEASE request in such a case.

We can determine if flush is being called with the last instance of file by
looking at the refcount.  But care is needed since the file descriptor
table is updated with RCU, so even if close() already cleared the
descriptor it could gain another reference from another CPU which sees the
old version of the table.  To prevent this, use synchronize_rcu() before
checking the refcount.

Note: this does not guarantee that the last close(2) will invoke a
synchronous release, since there may still be references to the file.  For
example:

 - being "in flight" with SCM_RIHTS in a unix domain socket
 - /proc/PID/fdinfo being accessed
 - multithreaded app doing parallel operation on the file descriptor being
   closed

In these cases RELEASE will be called from last fput() and will be
asynchronous, as before.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>