UPSTREAM: tty: serial: msm_serial: Use DT aliases
We rely on probe order of this driver to determine the line number for
the uart port. This makes it impossible to know the line number
when these devices are populated via DT. Use the DT alias
mechanism to assign the line based on the aliases node.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 97f754710304382607417c700476007443cd96a4)

BUG=chrome-os-partner:49255
TEST=Serial and DMA are working with these changes

Change-Id: Ic7c17d15d34018137626702635e52de4cd90bb13
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/331613
Commit-Ready: Matthew McClintock <mmcclint@qca.qualcomm.com>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>