UPSTREAM: tty: serial: msm: Reset uartdm after baud rate change
We need to issue a reset if we ever change the value of the IPR
register on DM hardware. If we don't reset the hardware the RX
stale interrupt never triggers and the only way to trigger an RX
handling event is by filling up the fifo. This causes things like
getty to not work so well considering it might change the baud
rate a few times. Fix this by moving the reset on startup and any
reprogramming required after the reset to be after we change the
baud rate.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit a12f1b406f2d3c3e3bba5300b6e420d004ca263b)

BUG=chrome-os-partner:49255
TEST=Serial and DMA are working with these changes

Change-Id: I02ce752b95fccffa07922aa33a4418591736f9c6
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/331612
Commit-Ready: Matthew McClintock <mmcclint@qca.qualcomm.com>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>