CHROMIUM: PM / wakeup: add tracking platform wakeup source
This adds an interface for platforms to track what caused the system to resume.
It logs the name of the wakeup source that woke the system, according to info
provided by the platform, and stores what type of wakeup it is. A device can be
classified as a user or automatic wakeup source through the wakeup sysfs
interface. This way, userspace can change its behavior based on what woke up the
system.

wakeup_type is added to the /sys/power path to tell what type of wakeup source
woke the system. The type of a wakeup source can be changed through its own
wakeup_type file located in its own power directory.

BUG=chromium:414949
TEST=emerge-samus chromeos-kernel-3_14 succeeds

Change-Id: I741ded8d1b6cf40d39965d3e70e71f68271c05f5
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/226932
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>

Conflicts:
	drivers/base/power/wakeup.c
	include/linux/pm.h
[rebase44(groeck): Resolved conflicts; squashed fixup patch]
Signed-off-by: Guenter Roeck <groeck@chromium.org>

Conflicts:
	drivers/base/power/wakeup.c
	include/linux/pm_wakeup.h
	kernel/power/main.c
[rebase412(groeck): Resolved conflicts; squashed with fixup commit]
Signed-off-by: Guenter Roeck <groeck@chromium.org>