ext4 crypto: add ext4 encryption facilities
We encrypt into bounce pages and schedule them for block I/O. We
decrypt in-place in the newly added read completion callback.

The current encryption mode, AES-256-XTS, is the first of several
encryption modes on the roadmap. In-plan modes include HMAC-SHA1 for
integrity-only and AES-256-GCM for authenticated encryption. These
future modes depend on anticipated functionality for storing per-block
metadata.

Signed-off-by: Michael Halcrow <mhalcrow@google.com>
Signed-off-by: Ildar Muslukhov <ildarm@google.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>