CHROMIUM: HID: usbhid: get/put around clearing needs_remote_wakeup
usbhid->intf->needs_remote_wakeup is set when a device is
opened, and is cleared when a device is closed.

When a usbhid device that does not support remote wake
( i.e. !device_can_wakeup() ) is closed, we fail out of
autosuspend_check() because the autosuspend check is called
before the flag is cleared as a result of usb_kill_urb(usbhid->urbin);

The result is that a device that may otherwise autosuspend will
fail to enter suspend again after all handles to it are closed.

In usbhid_open, usb_autopm_get_interface is called
before setting the needs_remote_wakeup flag, and
usb_autopm_put_interface is called after hid_start_in.

However, when the device is closed in usbhid_close, the same
protection isn't there when clearing needs_remote_wakeup. This will
add that to usbhid_close as well as usbhid_stop.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:431886
TEST=Whitelist security key devices using crosreview.com/229105
Plug in the security key. After a second, touch the device.
Verify that it has entered suspend (doesn't light up).
Bring up a 2 factor prompt that uses security key.
Touch the key to proceed through the prompt.
Check that it succeeds.
Check that after 2 seconds, the device returns to suspended state.

Change-Id: If462785d61b0e8f4d592ed064ce48d06c3e478b9
Reviewed-on: https://chromium-review.googlesource.com/229176
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Trybot-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>