UPSTREAM: dmaengine: Introduce a device_config callback
The fact that the channel configuration is done in device_control is rather
misleading, since it's not really advertised as such, plus, the fact that the
framework exposes a function of its own makes it not really intuitive, while
we're losing the type checking whenever we pass that unsigned long argument.

Add a device_config callback to dma_device, with a fallback on the old
behaviour for now for existing drivers to opt in.

Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
(cherry picked from commit 94a73e30dfe6722e9f4ef19f7892901d7d00eab1)

BUG=chrome-os-partner:49255
TEST=Serial and DMA are working with these changes

Change-Id: I07620a76564c9aa605d7750fe3db33274eae6794
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/332993
Commit-Ready: Matthew McClintock <mmcclint@qca.qualcomm.com>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>