FROMLIST: drm: Add Content Protection properties to drm
Add new standard connector properties to track whether content protection
(ex: hdcp) is desired by userspace. There are two properties involved,
"Content Protection" and "Content Protection KSV".

The "Content Protection" property allows userspace to request protection
on a connector. Set "Desired" to enable, "Undesired" to disable.

The "Content Protection KSV" property reflects the current state of
protection. If the KSV is 0, the connection is not protected. Once the
driver has enabled protection, it will update the the value with the KSV
(or similarly unique identifier, if not using HDCP) of the first-hop
device (sink or repeater).

Signed-off-by: Sean Paul <seanpaul@chromium.org>
[rebase49(groeck):
	Reapplied FROMLIST: https://patchwork.kernel.org/patch/5439871/
	Various structural changes to match upstream;
	combined with follow-up patch]
Signed-off-by: Guenter Roeck <groeck@chromium.org>

Change-Id: Iafd4a8b90b614f440a362f879ad5bc6944668daa